ai-course-code/chapter05_langchain_basics/intelligent_assistant/requirements.txt
# 智能对话助手依赖
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
